doc: update documentation
At one point this project was utilizing some example code from the cited
sources. This is now distinct enough that it can stand on it's own, at
least according to black duck.

As such, adopt CC0 1.0 Universal so that this project may be used as a
starting point without attribution or other strings.

Signed-off-by: Randolph Sapp <rs@ti.com>